Improve entry, add nodejs 7
Resolved the issues with user on entry
Handle home folder in a way that works on OSX
Note that mapping the home folder to the host home folder in this way causes
  the .bash** files to be used in the container as well. This required some
  small defensive mods to _those_ files.
Future will have to have some way to manage the home dot files safely.
Node7 is installed with tools via curl per the node website.
This required some rearranging of the RUN line to make that work.